Update for version 2.0.0, Code optimization, Start for MySQL Support
Moved the files and edited the package.
Added a dependencyhandler for handeling the worldguard, worldedit and
vault dependencies.
Moved the messages.yml file to the src dir to use it easier.
Added a filehandler for the config and messages file.
Cleaned up the landlord main class.
Moved the old database stuff to the deprecateddatabasehandler class.
This is TODO!